Add proper estimate of head position
Using the rider height (or estimating head length at 300mm if not
provided).

We were using 30mm (which is a bit of a pinhead!)

And tidy up the code a bit to use numpy vectors everywhere etc.

Also the weight distribution was a bit wrong I think because it was
using the end of the chainstay for the wheel centre, and thus excluding
the dropout.